rework on dashboard - turned the dashboard layout found on both portals to a component
