+++ booking/src/app/customer/layout.tsx
@@ -1,127 +1,25 @@
-'use client';
+'use server';
 
-import { useEffect, useState, useRef } from "react";
-import { FiMenu } from "react-icons/fi";
-import { FaHome, FaUser, FaCalendar } from "react-icons/fa";
-import { createClientSideClient } from "@/lib/supabase/CreateClientSideClient";
-import SideNavButton from "@/_components/SideNavButton";
-import { Button } from "@/_components/ui/Button";
-import { usePathname } from "next/navigation";
+import DashboardLayout from "@/_components/DashboardLayout";
+import SessionChecker from "@/_components/serverSide/SessionChecker";
+import { redirect } from "next/navigation";
 
-export default function DashboardLayout({ children }: { children: React.ReactNode }) {
-    const [expandSideNav, toggleSideNav] = useState(false);
-    const [isLg, setIsLg] = useState<null|boolean>(null);
-    const navRef = useRef<HTMLDivElement | null>(null);
+export default async function Layout({ children }: { children: React.ReactNode }) {
 
-    const pathname = usePathname();
+  const sessionCheck = await SessionChecker({ portal: "customer" });
 
-    function handleLogOut() {
-      const supabase = createClientSideClient();
-      supabase.auth.signOut().then(() => {
-        window.location.href = "/login";
-      });
-    }
-  
-    useEffect(() => {
-      if (typeof window === "undefined") return;
-  
-      const checkSize = () => {
-        if (window.innerWidth >= 1024) {
-          if (!isLg) {
-            toggleSideNav(false);
-            setIsLg(true);
-          }
-        } else {
-          if (isLg) {
-            setIsLg(false);
-          }
-        }
-      };
-  
-      window.addEventListener("resize", checkSize);
-      checkSize(); // Run once on mount
-  
-      return () => window.removeEventListener("resize", checkSize);
-    }, [isLg]);
-
-    useEffect(() => {
-      if (!expandSideNav) return;
-
-      const handleClickOutside = (event: MouseEvent) => {
-          if (navRef.current && !navRef.current.contains(event.target as Node)) {
-              toggleSideNav(false);
-          }
-      };
-
-      document.addEventListener("mousedown", handleClickOutside);
-      return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, [expandSideNav]);
-
-  
-  
-    return (
-      <div className="h-screen w-screen flex bg-[linear-gradient(to_top_right,#1a011f,#2b0231,#220127,#240132,#2e0249,#5b0d6d,#921f38,#7a1d31)]">
-        {/* Sidebar */}
-        <nav 
-          ref={navRef}
-          className={`h-full text-white p-4 space-y-4 shadow-lg cursor-pointer
-          ${expandSideNav ? 'w-64 shadow-2xl' : isLg ? 'w-64 shadow-2xl' : 'w-20 shadow-lg'}
-          transition-width duration-300 ease-in-out`}>
-  
-          <h2 className={`text-2xl font-bold mb-4 transition-opacity duration-300 ${expandSideNav ? 'opacity-100 block' : 'opacity-0 hidden'} lg:block lg:opacity-100`}>
-            Pawspace
-            <Button onClick={handleLogOut}>Log-out</Button>
-          </h2>
-  
-          <SideNavButton 
-            icon={FiMenu} 
-            text="" 
-            onClick={() => toggleSideNav(!expandSideNav)}
-            color="purple"
-            className={`lg:hidden ${expandSideNav ? 'hidden' : ''}`}
-            isCurrent={false}
-          />
-  
-          <ul className="space-y-4">
-            {/* Add side navigation buttons here. Use SideNavButton Component */}
-            <SideNavButton icon={FaHome} 
-              text="Home" 
-              isCurrent={pathname === "/customer"} 
-              showText={expandSideNav} 
-              href="/customer"
-              color="purple"
-              onClick={() => {toggleSideNav(false)}}
-            />
-
-            <SideNavButton 
-              icon={FaCalendar} 
-              text="History" 
-              isCurrent={pathname === "/customer/history"} 
-              showText={expandSideNav} 
-              href="/customer/history"
-              color="purple"
-              onClick={() => {toggleSideNav(false)}}
-            />
-
-            <SideNavButton 
-              icon={FaUser} 
-              text="About Us" 
-              isCurrent={pathname === "/customer/about-us"} 
-              showText={expandSideNav} 
-              href="/customer/about-us"
-              color="purple"
-              onClick={() => {toggleSideNav(false)}}
-            />
-
-          </ul>
-        </nav>
-  
-        {/* Main Content */}
-        <div className={`flex-1 p-6 transition-all duration-300 ease-in-out overflow-x-hidden overflow-y-auto ${expandSideNav ? 'opacity-20' : 'opacity-100'}`}>
-          {children}
-        </div>
-        
-      </div>
-    );
+  if(!sessionCheck) {
+    redirect("/login");
   }
-  
+
+  return (
+    <DashboardLayout colorTheme="purple"
+    buttons={[
+      {icon: "FaHome", text: "Home", href: "/customer"},
+      {icon: "Facalendar", text: "History", href: "/customer/history"},
+      {icon: "FaUser", text: "About Us", href: "/customer/about-us"},
+    ]}>
+      {children}
+    </DashboardLayout>
+  );
+}
